Copy tests between Single/Double.

+++ src/ConvNetSharp.Volume/Single/Volume.cs
@@ -187,7 +187,7 @@ namespace ConvNetSharp.Volume.Single
             for (var n = 0; n < batchSize; n++)
             {
                 // compute max activation
-                var amax = double.MinValue;
+                var amax = float.MinValue;
                 for (var depth = 0; depth < outputDepth; depth++)
                 {
                     for (var ay = 0; ay < outputHeight; ay++)
@@ -204,8 +204,8 @@ namespace ConvNetSharp.Volume.Single
                 }
 
                 // compute exponentials (carefully to not blow up)
-                var es = new double[outputDepth * outputHeight * outputWidth];
-                var esum = 0.0;
+                var es = new float[outputDepth * outputHeight * outputWidth];
+                var esum = 0.0f;
 
                 for (var depth = 0; depth < outputDepth; depth++)
                 {
@@ -213,7 +213,7 @@ namespace ConvNetSharp.Volume.Single
                     {
                         for (var ax = 0; ax < outputWidth; ax++)
                         {
-                            var e = Math.Exp(Get(ax, ay, depth, n) - amax);
+                            var e = (float)Math.Exp(Get(ax, ay, depth, n) - amax);
                             esum += e;
                             es[ax + ay * outputWidth + depth * outputWidth * outputHeight] = e;
                         }
@@ -228,9 +228,7 @@ namespace ConvNetSharp.Volume.Single
                         for (var ax = 0; ax < outputWidth; ax++)
                         {
                             es[ax + ay * outputWidth + depth * outputWidth * outputHeight] /= esum;
-
-                            result.Storage.Set(ax, ay, depth, n,
-                                (float)es[ax + ay * outputWidth + depth * outputWidth * outputHeight]);
+                            result.Storage.Set(ax, ay, depth, n, es[ax + ay * outputWidth + depth * outputWidth * outputHeight]);
                         }
                     }
                 }
@@ -243,7 +241,7 @@ namespace ConvNetSharp.Volume.Single
         }
 
         public override void DoPool(Volume<float> result, int windowWidth, int windowHeight,
-                   int horizontalPad, int verticalPad, int horizontalStride, int verticalStride)
+            int horizontalPad, int verticalPad, int horizontalStride, int verticalStride)
         {
             var inputWidth = this.Shape.GetDimension(0);
             var inputHeight = this.Shape.GetDimension(1);
